partially working, vm is busted still
Hard to say what exactly is going wrong, but `tests/vm/tnimnode.nim`
sigsegvs within a when expression for `arithmetics.ze`.

This is likely because reference handling is broken given the current
approach and there isn't enough information or tests to spot the errors.
Added a pile of asserts but that hasn't helped, it's almost certainly
in the VM, however.

Need a new debugging/test strategy.