saving things, thought of a new approach
could this ref/deref thing be as simple as whenever we take a ref
(`addr`), just create a nkIntLit node with an intVal set to the NodeId
of the thing we took an `addr` of and whenever we deref (`[]`), return
the specified NodeId? To tell them apart `rkNode` vs `rkNodeAddr`
and don't screw up vmgen.